chatbot/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
babel==2.16.0
bibtexparser==2.0.0b8
certifi==2024.12.14
charset-normalizer==3.4.1
clldutils==3.24.0
cohere==5.13.5
colorama==0.4.6
colorlog==6.9.0
csvw==3.5.1
datasets==3.2.0
dill==0.3.8
dlinfo==2.0.0
et_xmlfile==2.0.0
faiss-cpu==1.9.0.post1
Faker==33.3.0
fastavro==1.10.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
isodate==0.7.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
language-tags==1.2.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
np==1.0.2
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
phonemizer==3.3.0
pillow==11.1.0
propcache==0.2.1
pyarrow==19.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pylatexenc==2.10
PyMySQL==1.1.1
pyodbc==5.2.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rdflib==7.1.3
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.15.0
segments==2.2.1
sentence-transformers==3.4.0
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
Flask==3.1.0